Moved asset compoenets to AssetBoard dir.' Made space for ShoppingCart
