+++ client/src/route/library/Library.tsx
@@ -1,12 +1,11 @@
 import * as React from 'react';
 import Layout from 'page/Layout';
 import TabComponent from 'components/tab/TabComponent';
-import Iframe from 'components/Iframe';
 import { TabData } from 'components/tab/subcomponents/TabRender';
-import { useURLforLIB } from 'util/envUtil';
-import { Typography } from '@mui/material';
-import { CardData } from 'components/card/AssetCard';
-import AssetBoard from 'components/card/AssetBoard';
+// import { useURLforLIB } from 'util/envUtil';
+import { Paper, Typography } from '@mui/material';
+import { CardData } from 'components/AssetBoard/AssetCard';
+import AssetBoard from 'components/AssetBoard';
 import tabs from './LibraryTabData';
 
 const cardData: CardData[] = [
@@ -52,17 +51,14 @@ const cardData: CardData[] = [
 ];
 
 function useLibraryData() {
-  const LIBurl = useURLforLIB();
+  // const LIBurl = useURLforLIB();
 
   const tabsData: TabData[] = tabs.map((tab) => ({
     label: tab.label,
     body: (
       <>
         <Typography variant="body1">{tab.body}</Typography>
-        {tab.label === 'Functions' && <AssetBoard data={cardData} />}
-        {tab.label !== 'Functions' && (
-          <Iframe title={`JupyterLight-Demo-${tab.label}`} url={LIBurl} />
-        )}
+        <AssetBoard data={cardData} />
       </>
     ),
   }));
@@ -73,7 +69,22 @@ function LibraryContent() {
   const tabsData = useLibraryData();
   return (
     <Layout>
-      <TabComponent tabs={tabsData} />
+      <div style={{ display: 'flex' }}>
+        <TabComponent
+          tabs={tabsData}
+          sx={{ flexGrow: 2, marginRight: '2rem' }}
+        />
+        <Paper
+          sx={{
+            flexGrow: 1,
+            minWidth: '20rem',
+            textAlign: 'center',
+            paddingTop: '2rem',
+          }}
+        >
+          ShoppingCart
+        </Paper>
+      </div>
     </Layout>
   );
 }
